tts/requirements.txt
azure-cognitiveservices-speech
google-cloud-texttospeech
pydub
pydantic
fastapi
uvicorn